requirements.txt
# Core Flask packages
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # For PostgreSQL

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# ML Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# NLP
nltk==3.8.1

# Time Series Forecasting (optional)
prophet==1.1.4

# Anomaly Detection
scipy==1.11.2

# API calls
requests==2.31.0

# Data processing
python-dateutil==2.8.2

# Environment management
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Additional packages for better functionality
redis==4.6.0
celery==5.3.1
marshmallow==3.20.1
flask-marshmallow==0.15.0